celeryr/requirements.txt
celery==5.2.3
flower==1.0.0
redis==4.1.4
psycopg2==2.9.6
transformers
llama-cpp-python
newsdataapi
torch
huggingface_hub
beautifulsoup4==4.12.2
SQLAlchemy==2.0.16
